requirements.txt
numpy=1.26.3
pandas=2.2.1
pytorch=2.1.2
tqdm=4.66.2